change type long to int in all methods 'sum'

+++ src/main/java/com/epam/cdp/m2/hw2/aggregator/Aggregator.java
@@ -13,7 +13,7 @@ public interface Aggregator {
      * @param numbers list of integers to sum up
      * @return sum
      */
-    long sum(List<Integer> numbers); // simple method just to warm up :)
+    int sum(List<Integer> numbers); // simple method just to warm up :)
 
     /**
      * Use this method to find words frequency.
